refactor: move check threshold and send alert function
